refactory : 폴더 구조 및 mapping 변경
