+++ src/main/java/com/agos/awg/controller/CardController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.agos.awg.model.CardVO;
-import com.agos.awg.service.CardService;
+import com.agos.awg.service.card.CardService;
 
 @Controller
 public class CardController {
@@ -22,27 +22,27 @@ public class CardController {
 	// 카드 목록 페이지 이동
 	@RequestMapping("/card")
 	public String cardInfo(Model model) {
-		return "cardInfo";
+		return "card/cardInfo";
 	}
 	
 	// 카드 목록 리스트(카드 유형별)
-	@RequestMapping("/cardList")
+	@RequestMapping("/card/cardList")
 	public String cardList(@RequestParam("card_type") int card_type, Model model) {
 		ArrayList<CardVO> cardList = service.getCardList(card_type);
 		
 		model.addAttribute("cardList", cardList);
-		return "cardList";
+		return "card/cardList";
 	}
 	
 	// 카드 계정 등록 페이지 
-	@RequestMapping("/insertCardView")
+	@RequestMapping("/card/insertCardView")
 	public String insertCardView() {
-		return "insertCard";
+		return "card/insertCard";
 	}
 	
 	// 카드 계정 중복 체크
 	@ResponseBody
-	@RequestMapping("/checkCardIndex")
+	@RequestMapping("/card/checkCardIndex")
 	public int checkCardIndex(@RequestParam("index") String index) {
 		int result = service.checkCardIndex(index);		
 		return result;
@@ -50,7 +50,7 @@ public class CardController {
 	
 	// 카드 계정 등록
 	@ResponseBody
-	@RequestMapping("/insertCard")
+	@RequestMapping("/card/insertCard")
 	public String insertCard(CardVO cardVO) {
 		service.insertCard(cardVO);
 		
@@ -59,17 +59,17 @@ public class CardController {
 	}
 	
 	// 카드 계정 수정 페이지 
-	@RequestMapping("/updateCardView/{index}")
+	@RequestMapping("/card/updateCardView/{index}")
 	public String updateCardView(@PathVariable("index") String card_idx, Model model) {
 		CardVO card = service.getCardInfo(card_idx);
 		
 		model.addAttribute("card", card);
-		return "updateCard";
+		return "card/updateCard";
 	}
 	
 	// 카드 정보 수정
 	@ResponseBody
-	@RequestMapping("/updateCard")
+	@RequestMapping("/card/updateCard")
 	public int updateCard(CardVO cardVO) {
 		int result = service.updateCard(cardVO);
 		
@@ -78,7 +78,7 @@ public class CardController {
 	
 	// 카드 정보 삭제
 	@ResponseBody
-	@RequestMapping("/deleteCard")
+	@RequestMapping("/card/deleteCard")
 	public int deleteCard(@RequestParam("indexArray[]") List<String> indexArray) {
 		int result = service.deleteCard(indexArray);
