+++ src/main/java/com/agos/awg/service/card/CardService.java
@@ -1,4 +1,4 @@
-package com.agos.awg.service;
+package com.agos.awg.service.card;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,7 +6,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.agos.awg.dao.ICardDAO;
+import com.agos.awg.dao.card.ICardDAO;
 import com.agos.awg.model.CardVO;
 
 @Service
