+++ src/main/webapp/resources/css/card/cardInfo.css
